Applied additions/changes to Restaurant and Item controllers.

+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/entity/OrdersEntity.java
@@ -27,14 +27,14 @@ public class OrdersEntity implements Serializable {
 
     @Column(name = "bill")
     @NotNull
-    private Integer bill;
+    private Double bill;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "coupon_id")
     private CouponEntity couponId;
 
     @Column(name = "discount")
-    private Integer discount;
+    private Double discount;
 
     @Column(name = "date")
     private Timestamp date;
@@ -77,19 +77,19 @@ public class OrdersEntity implements Serializable {
         this.uuid = uuid;
     }
 
-    public Integer getBill() {
+    public Double getBill() {
         return bill;
     }
 
-    public void setBill(Integer bill) {
+    public void setBill(Double bill) {
         this.bill = bill;
     }
 
-    public Integer getDiscount() {
+    public Double getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Integer discount) {
+    public void setDiscount(Double discount) {
         this.discount = discount;
     }
